+++ src/events/IndexAssetEvent.php
@@ -1,0 +1,11 @@
+<?php
+namespace xorb\search\events;
+
+use craft\base\Event;
+use craft\elements\Asset;
+
+class IndexAssetEvent extends Event
+{
+    public Asset $asset;
+    public ?string $mainData;
+}