+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.1.0 - Unreleased
+
+### Added
+
+- Added PageResult::EVENT\_UPDATE\_MAIN\_DATA event to allow custom indexing of pages.
+- Added AssetResult::EVENT\_UPDATE\_MAIN\_DATA event to allow custom indexing of assets.
+- Added option to force update pages and assets so that custom implementations can be run on pages and assets that have not changed.
+
+### Fixed
+
+- Fixed issue with searchable asset field value not holding in certain situations.
+
 ## 1.0.0 - 2024-07-27
 
 Initial release.